some configs mvc tests

+++ src/main/java/ru/bbpax/keeper/model/Ingredient.java
@@ -7,6 +7,8 @@ import org.bson.types.ObjectId;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Field;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @QueryEmbeddable
@@ -15,10 +17,10 @@ public class Ingredient {
     @Field("_id")
     private String id;
     private String name;
-    private float value;
+    private BigDecimal value;
     private String unit;
 
-    public Ingredient(String name, float value, String unit) {
+    public Ingredient(String name, BigDecimal value, String unit) {
         this();
         this.name = name;
         this.value = value;
